server/src/requirements.txt
influxdb-client==1.48.0
ciso8601==2.3.2
Flask>=3.1.0
python-dotenv>=1.1.0